mcp-servers/google_workspace/requirements.txt
mcp>=1.8.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-cloud-firestore>=2.16.0
uvicorn>=0.30.0
starlette>=0.37.0
python-dotenv>=1.0.0
